Elimination

StructuralIdentifiability.Bezout_matrix

StructuralIdentifiability.Sylvester_matrix

StructuralIdentifiability.choose

StructuralIdentifiability.eliminate_var

StructuralIdentifiability.simplify_matrix

StructuralIdentifiability.Bezout_matrix — Method

Bezout_matrix(f, g, var_elim)

Compute the Bezout matrix of two polynomials f, g with respect to var_elim

Inputs:

f - first polynomial

g - second polynomial

var_elim - variable, of which f and g are considered as polynomials

Output:

M::MatrixElem - The Bezout matrix

StructuralIdentifiability.Sylvester_matrix — Method

Sylvester_matrix(f, g, var_elim)

Compute the Sylvester matrix of two polynomials f, g with respect to var_elim Inputs:

f - first polynomial

g - second polynomial

var_elim - variable, of which f and g are considered as polynomials

Output:

M::MatrixElem - The Sylvester matrix

StructuralIdentifiability.choose — Method

choose(polys, generic_point_generator)

Input:

polys - an array of distinct irreducible polynomials in the same ring

generic_point_generator - a generic point generator as described above for one of polys

Output:

the polynomial that vanishes at the generic_point_generator

StructuralIdentifiability.eliminate_var — Method

eliminate_var(f, g, var_elim, generic_point_generator)

Eliminate variable from a pair of polynomials

Input:

f and g - polynomials

var_elim - variable to be eliminated

generic_point_generator - a generic point generator object for the factor of the resultant of f and g of interest

Output:

polynomial - the desired factor of the resultant of f and g

StructuralIdentifiability.simplify_matrix — Method

simplify_matrix(M)

Eliminate GCD of entries of every row and column

Input:

M::MatrixElem - matrix to be simplified

Output:

M::MatrixElem - Simplified matrix

extra_factors::Vector{AbstractAlgebra.MPolyElem} - array of GCDs eliminated from M.