Global Identifiability Tools

StructuralIdentifiability.check_field_membership

StructuralIdentifiability.extract_identifiable_functions

StructuralIdentifiability.find_identifiable_functions

StructuralIdentifiability.get_degree_and_coeffsize

StructuralIdentifiability.extract_identifiable_functions — Function

extract_identifiable_functions(io_equations, parameters)

For the io_equation and the list of all parameter variables, returns a set of generators of a field of all functions of parameters

Note: an experimental functionality at the moment, may fail and may be inefficient

StructuralIdentifiability.check_field_membership — Function

check_field_membership(generators, rat_funcs, p, [method=:GroebnerBasis])

Checks whether given rational function belong to a given field of rational functions

Inputs:

generators - a list of lists of polynomials. Each of the lists, say, [f1, ..., fn], defines generators f2/f1, ..., fn/f1. Let $F$ be the field generated by all of them.

rat_funcs - list of rational functions

p - a real number between 0 and 1, the probability of correctness

Output:

a list L[i] of bools of length length(rat_funcs) such that L[i] is true iff the i-th function belongs to $F$. The whole result is correct with probability at least p

StructuralIdentifiability.find_identifiable_functions — Function

find_identifiable_functions(ode::ODE{<: MPolyElem{fmpq}}, p::Float64=0.99)

Input:

ode - ODE-system

p - probability of correctness

Output:

returns a set of generators of the field of all functions of parameters

Find identifiable functions of parameters for a given ode.

StructuralIdentifiability.get_degree_and_coeffsize — Function

get_degree_and_coeffsize(f)

for f being a polynomial/rational function over rationals (QQ) returns a tuple (degree, max_coef_size)