Local Identifiability Tools

StructuralIdentifiability.differentiate_output

StructuralIdentifiability.differentiate_solution

StructuralIdentifiability.differentiate_solution — Function

differentiate_solution(ode, params, ic, inputs, prec)

Input:

the same as for power_series_solutions

Output:

a tuple consisting of the power series solution and a dictionary of the form (u, v) => power series, where u is a state variable v is a state or parameter, and the power series is the partial derivative of the function u w.r.t. v evaluated at the solution

StructuralIdentifiability.differentiate_output — Function

differentiate_output(ode, params, ic, inputs, prec)

Similar to differentiate_solution but computes partial derivatives of prescribed outputs returns a dictionary of the form y_function => Dict(var => dy/dvar) where dy/dvar is the derivative of y_function with respect to var.