Working with ODEs

StructuralIdentifiability.PreprocessODE

StructuralIdentifiability._reduce_mod_p

StructuralIdentifiability.power_series_solution

StructuralIdentifiability.reduce_ode_mod_p

StructuralIdentifiability.set_parameter_values

StructuralIdentifiability.PreprocessODE — Method

function PreprocessODE(de::ModelingToolkit.ODESystem, measured_quantities::Array{ModelingToolkit.Equation})

Input:

de - ModelingToolkit.ODESystem, a system for identifiability query

measured_quantities - array of output functions

Output:

ODE object containing required data for identifiability assessment

StructuralIdentifiability._reduce_mod_p — Method

_reduce_mod_p(f, p)

Reduces a polynomial/rational function over Q modulo p

StructuralIdentifiability.power_series_solution — Method

power_series_solution(ode, param_values, initial_conditions, input_values, prec)

Input:

ode - an ode to solve

param_values - parameter values, must be a dictionary mapping parameter to a value

initial_conditions - initial conditions of ode, must be a dictionary mapping state variable to a value

input_values - power series for the inpiuts presented as a dictionary variable => list of coefficients

prec - the precision of solutions

Output:

computes a power series solution with precision prec presented as a dictionary variable => corresponding coordiante of the solution

StructuralIdentifiability.reduce_ode_mod_p — Method

reduce_ode_mod_p(ode, p)

Input: ode is an ODE over QQ, p is a prime number Output: the reduction mod p, throws an exception if p divides one of the denominators

StructuralIdentifiability.set_parameter_values — Method

set_parameter_values(ode, param_values)

Input:

ode - an ODE as above

param_values - values for (possibly, some of) the parameters as dictionary parameter => value

Output:

new ode with the parameters in param_values plugged with the given numbers