Power Series Utilities

StructuralIdentifiability._matrix_inv_newton_iteration

StructuralIdentifiability.ps_diff

StructuralIdentifiability.ps_integrate

StructuralIdentifiability.ps_matrix_homlinear_de

StructuralIdentifiability.ps_matrix_inv

StructuralIdentifiability.ps_matrix_linear_de

StructuralIdentifiability.ps_matrix_log

StructuralIdentifiability.ps_ode_solution

StructuralIdentifiability._matrix_inv_newton_iteration — Method

_matrix_inv_newton_iteration(M, Minv)

Performs a single step of Newton iteration for inverting M with Minv being a partial result

StructuralIdentifiability.ps_diff — Method

ps_diff(ps)

Input:

ps - (absolute capped) unvariate power series

Output:

the derivative of ps

StructuralIdentifiability.ps_integrate — Method

ps_integrate(ps)

Input:

ps - (absolute capped) unvariate power series

Output:

the integral of ps without constant term

StructuralIdentifiability.ps_matrix_homlinear_de — Method

ps_matrix_homlinear_de(A, Y0, prec)

Input:

A - a square matrix with entries in a univariate power series ring

Y0 - a square invertible matrix over the base field

Output:

matrix Y such that Y' = AY up to precision of A - 1 and Y(0) = Y0

StructuralIdentifiability.ps_matrix_inv — Function

ps_matrix_inv(M, prec)

Input:

M - a square matrix with entries in a univariate power series ring it is assumed that M(0) is invertible and all entries having the same precision

prec - an integer, precision, if -1 then defaults to precision of M

Output:

the inverse of M computed up to prec

StructuralIdentifiability.ps_matrix_linear_de — Method

ps_matrix_linear_de(A, B, Y0, prec)

Input:

A, B - square matrices with entries in a univariate power series ring

Y0 - a matrix over the base field with the rows number the same as A

Output:

matrix Y such that Y' = AY + B up to precision of A - 1 and Y(0) = Y0

StructuralIdentifiability.ps_matrix_log — Method

ps_matrix_log(M)

Input:

M - a square matrix with entries in a univariate power series ring it is assumed that M(0) is the identity

Output:

the natural log of M

StructuralIdentifiability.ps_ode_solution — Method

ps_ode_solution(equations, ic, inputs, prec)

Input:

equations - a system of the form $A(x, u, mu)x' - B(x, u, mu) = 0$, where A is a generically nonsingular square matrix. Assumption: A is nonzero at zero

ic - initial conditions for x's (dictionary)

inputs - power series for inputs represented as arrays (dictionary)

prec - precision of the solution

Output:

power series solution of the system