Global Identifiability Tools

StructuralIdentifiability.RationalFunctionField

StructuralIdentifiability.field_contains

StructuralIdentifiability.get_degree_and_coeffsize

StructuralIdentifiability.RationalFunctionField — Type

RationalFunctionField

A subfield of the field of rational functions over the rationals.

Example

using Nemo
using StructuralIdentifiability: RationalFunctionField

R, (x, y, z) = QQ["x", "y", "z"]

# Constructs a subfield generated by x / y, y / z
rff = RationalFunctionField([x // y, y // z])

# Constructs a subfield generated by y / x, 1 / x, z / y
rff = RationalFunctionField([[x, y, R(1)], [y, z]])

StructuralIdentifiability.field_contains — Function

field_contains(field, ratfuncs, prob_threshold)

Checks whether given rational function field field contains given rational functions ratfuncs (represented as a list of lists). The result is correct with probability at least prob_threshold

Inputs:

field - a rational function field

ratfuncs - a list of lists of polynomials. Each of the lists, say, [f1, ..., fn], defines generators f2/f1, ..., fn/f1.

prob_threshold real number from (0, 1)

Output:

a list L[i] of bools of length length(rat_funcs) such that L[i] is true iff the i-th function belongs to field

StructuralIdentifiability.get_degree_and_coeffsize — Function

get_degree_and_coeffsize(f)

for f being a polynomial/rational function over rationals (QQ) returns a tuple (degree, max_coef_size)