Functions to work with the ODE structure

StructuralIdentifiability.__preprocess_ode — Method

function __preprocess_ode(de::ModelingToolkit.AbstractTimeDependentSystem, measured_quantities::Array{Tuple{String, SymbolicUtils.BasicSymbolic}})

Input:

de - ModelingToolkit.AbstractTimeDependentSystem, a system for identifiability query

measured_quantities - array of input function in the form (name, expression)

Output:

ODE object containing required data for identifiability assessment

conversion dictionary from the symbols in the input MTK model to the variable involved in the produced ODE object

StructuralIdentifiability._get_var — Method

For an expression of the form f'(t) or f(t) returns (f, true) and (f, false), resp

StructuralIdentifiability._reduce_mod_p — Method

_reduce_mod_p(f, p)

Reduces a polynomial/rational function over Q modulo p

StructuralIdentifiability.power_series_solution — Method

power_series_solution(ode, param_values, initial_conditions, input_values, prec)

Input:

ode - an ode to solve

param_values - parameter values, must be a dictionary mapping parameter to a value

initial_conditions - initial conditions of ode, must be a dictionary mapping state variable to a value

input_values - power series for the inputs presented as a dictionary variable => list of coefficients

prec - the precision of solutions

Output:

computes a power series solution with precision prec presented as a dictionary variable => corresponding coordinate of the solution

StructuralIdentifiability.preprocess_ode — Method

function preprocess_ode(de::ModelingToolkit.AbstractTimeDependentSystem, measured_quantities::Array{ModelingToolkit.Equation})
function preprocess_ode(de::ModelingToolkit.AbstractTimeDependentSystem, measured_quantities::Array{SymbolicUtils.BasicSymbolic})

Input:

de - ModelingToolkit.AbstractTimeDependentSystem, a system for identifiability query

measured_quantities - array of output functions (as equations of just functions)

Output:

ODE object containing required data for identifiability assessment

conversion dictionary from the symbols in the input MTK model to the variable involved in the produced ODE object

StructuralIdentifiability.reduce_ode_mod_p — Method

reduce_ode_mod_p(ode, p)

Input: ode is an ODE over QQ, p is a prime number Output: the reduction mod p, throws an exception if p divides one of the denominators

StructuralIdentifiability.set_parameter_values — Method

set_parameter_values(ode, param_values)

Input:

ode - an ODE as above

param_values - values for (possibly, some of) the parameters as dictionary parameter => value

Output:

new ode with the parameters in param_values plugged with the given numbers

StructuralIdentifiability.find_submodels — Method

find_submodels(ode)

The function calculates and returns all valid submodels given a system of ODEs.

Input:

ode - an ODEs system to be studied

Output:

A list of submodels represented as ode objects

Example:

>ode = @ODEmodel(x1'(t) = x1(t)^2, 
                 x2'(t) = x1(t) * x2(t), 
                 y1(t) = x1(t), 
                 y2(t) = x2(t))
>find_submodels(ode)
    ODE{fmpq_mpoly}[
        
        x1'(t) = a(t)*x2(t)^2 + x1(t)
        y1(t) = x1(t)
    ]