Other Helpful Functions

StructuralIdentifiability.decompose_derivative

StructuralIdentifiability.extract_coefficients

StructuralIdentifiability.make_substitution

StructuralIdentifiability.replace_with_ic

StructuralIdentifiability.switch_ring

StructuralIdentifiability.uncertain_factorization

StructuralIdentifiability.decompose_derivative — Method

decompose_derivative(varname, prefixes)

Determines if it is possible to represent the varname as a_number where a is an element of prefixes If yes, returns a pair (a, number), otherwise nothing

StructuralIdentifiability.extract_coefficients — Method

extract_coefficients(poly, variables)

Input:

poly - multivariate polynomial

variables - a list of variables from the generators of the ring of poly

Output:

dictionary with keys being tuples of length length(variables) and values being polynomials in the variables other than those which are the coefficients at the corresponding monomials (in a smaller polynomial ring)

StructuralIdentifiability.make_substitution — Method

make_substitution(f, var_sub, val_numer, val_denom)

Substitute a variable in a polynomial with an expression

Input:

f - the polynomial

var_sub - the variable to be substituted

var_numer - numerator of the substitution expression

var_denom - denominator of the substitution expression

Output:

polynomial - result of substitution

StructuralIdentifiability.replace_with_ic — Method

replace_with_ic(ode::ODE, funcs)

Takes an ode and a list of functions in the states and parameters and makes a change of variable names x(t) -> x(0). Function is used to prepare the output for the case of known initial conditions

StructuralIdentifiability.switch_ring — Method

switch_ring(v, ring)

For a variable v, returns a variable in ring with the same name

StructuralIdentifiability.uncertain_factorization — Method

uncertain_factorization(f)

Input:

f - polynomial with rational coefficients

Output:

list of pairs (div, certainty) where

div's are divisors of f such that f is their product with certain powers

if certainty is true, div is $Q$-irreducible