Wronskian Tools

StructuralIdentifiability.get_max_below — Method

get_max_below(t, vect)

Input:

t - a trie with exponent vectors

vect - yet another exponent vector

Output:

a pair (d, v) where v is a vector in the trie which is componentwise ≤ vect and the difference d is as small as possible

StructuralIdentifiability.massive_eval — Method

massive_eval(polys, eval_dict)

Input:

polys - a list of polynomials

eval_dict - dictionary from variables to the values. Missing values are treated as zeroes

Output:

a list of values of the polynomials

Evaluates a list of polynomials at a point. Assumes that multiplications are relatively expensive (like in truncated power series) so all the monomials are precomputed first and the values of monomials of lower degree are cached and used to compute the values of the monomials of higher degree

StructuralIdentifiability.monomial_compress — Method

monomial_compress(io_equation, ode)

Compresses an input-output equation for the rank computation Input:

io_equation - input-output equation

ode - the corresponding ODE model

Output:

pair (coeffs, terms) such that:

sum of coeffs[i] * terms[i] = io_equation

coeffs involve only parameters, terms involve only inputs and outputs

length of the representation is the smallest possible

StructuralIdentifiability.wronskian — Method

wronskian(io_equations, ode)

Input:

io_equations - a set of io-equations in the form of the Dict as returned by find_ioequations

ode - the ODE object

Output:

a list of Wronskians evaluated at a point modulo prime

Computes the Wronskians of io_equations